Struct uart_pl011::Pl011

pub struct Pl011 { /* private fields */ }

A Pl011 Single-Serial-Port Controller.

Implementations

impl Pl011

Generic methods

pub fn new(base: PhysicalAddress) -> Result<Self, &'static str>

Initialize a UART driver.

pub fn enable_rx_interrupt(&mut self, enable: bool)

Enable on-receive interrupt

pub fn acknowledge_rx_interrupt(&mut self)

pub fn set_fifo_mode(&mut self, enable: bool)

Set FIFO mode

pub fn log_status(&self)

Outputs a summary of the state of the controller using log::info!()

pub fn has_incoming_data(&self) -> bool

Returns true if the receive-buffer-empty flag is clear.

pub fn read_byte(&self) -> u8

Reads a single byte out the uart

Spins until a byte is available in the fifo.

pub fn read_bytes(&self, bytes: &mut [u8]) -> usize

Reads bytes into a slice until there is none available.

pub fn is_writeable(&self) -> bool

Returns true if the transmit-buffer-full flag is clear.

pub fn write_byte(&mut self, data: u8)

Writes a single byte out the uart.

Spins until space is available in the fifo.

pub fn write_bytes(&mut self, bytes: &[u8])

Writes a byte slice out the uart.

Spins until space is available in the fifo.